unsupervised/dimensionality-reduction/UMAP/requirements.txt
scikit-learn
matplotlib
umap-learn